+++ src/machine_learning/classes/networks/multilayer_perceptron/Neuron.java
@@ -14,14 +14,14 @@ public class Neuron {
     public static final int EXTRA_WEIGHTS = 2;
     protected final int START_POINT, END_POINT, ADD_BIAS_POS, MUL_BIAS_POS;
     protected final Function function;
-    protected double[] allWeights;
+    protected float[] allWeights;
     private double neuronSum;
 
-    public Neuron(double[] allWeights, int startPoint, int endPoint, int maxStartValue) {
+    public Neuron(float[] allWeights, int startPoint, int endPoint, int maxStartValue) {
         this(allWeights, startPoint, endPoint, maxStartValue, null);
     }
 
-    public Neuron(double[] allWeights, int startPoint, int endPoint, int maxStartValue, Function function) {
+    public Neuron(float[] allWeights, int startPoint, int endPoint, int maxStartValue, Function function) {
         if (allWeights == null) {
             throw new IllegalArgumentException(
                     "Arguments, at pos 1: not null."
@@ -36,7 +36,7 @@ public class Neuron {
         MUL_BIAS_POS = endPoint - 2;
 
         for (int i = startPoint; i < endPoint; i++) {
-            allWeights[i] = Math.random() * (maxStartValue * 2) - maxStartValue;
+            allWeights[i] = (float) (Math.random() * (maxStartValue * 2) - maxStartValue);
         }
 
         this.function = function;
@@ -54,7 +54,7 @@ public class Neuron {
         return neuronSum;
     }
 
-    protected void setWeightAt(double weight, int position) {
+    protected void setWeightAt(float weight, int position) {
         if (position >= END_POINT - START_POINT - EXTRA_WEIGHTS || position < 0) throw new RuntimeException(
                 "position out of bounds. position=" + position + " length=" + (END_POINT - START_POINT)
         );
@@ -70,7 +70,7 @@ public class Neuron {
         return allWeights[START_POINT + position];
     }
 
-    protected void addToWeight(double value, int position) {
+    protected void addToWeight(float value, int position) {
         if (position >= END_POINT - START_POINT - EXTRA_WEIGHTS || position < 0) throw new RuntimeException(
                 "position out of bounds. position=" + position + " length=" + (END_POINT - START_POINT)
         );
@@ -78,7 +78,7 @@ public class Neuron {
         allWeights[START_POINT + position] += value;
     }
 
-    protected void setBias(double newBias) {
+    protected void setBias(float newBias) {
         allWeights[ADD_BIAS_POS] = newBias;
     }
 
@@ -86,18 +86,18 @@ public class Neuron {
         return allWeights[ADD_BIAS_POS];
     }
 
-    protected void addToBias(double value) {
+    protected void addToBias(float value) {
         allWeights[ADD_BIAS_POS] += value;
     }
 
-    protected double compute(double[] features) {
+    protected float compute(float[] features) {
         if (features.length != END_POINT - START_POINT - EXTRA_WEIGHTS) {
             throw new IllegalArgumentException(
                     "Features length is different from weights length, features=" + features.length + " weights=" + (END_POINT - START_POINT)
             );
         }
 
-        double sum = 0;
+        float sum = 0;
         if (function == null) {
             for (int i = START_POINT; i < END_POINT - EXTRA_WEIGHTS; i++) {
                 sum += allWeights[i] * features[i - START_POINT];
