+++ src/neuroevolution/classes/neuroevolution/Genes.java
@@ -12,27 +12,27 @@ import evolution_builder.population.Person;
  *
  * @author Zachs
  */
-class Genes implements evolution_builder.population.Genes<Double>{
+class Genes implements evolution_builder.population.Genes<Float>{
     static int maxStartValue = 1;
     
     @Override
-    public Double getGenAt(Person person, int position) {
+    public Float getGenAt(Person person, int position) {
         return ((Network) person.getGeneCode()).getWeightAt(position);
     }
 
     @Override
-    public void setGenAt(Person person, Double gene, int position) {
+    public void setGenAt(Person person, Float gene, int position) {
         ((Network) person.getGeneCode()).setWeightAt(position, gene);
     }
 
     @Override
     public void mutationValue(Person person, int position, double mutationValue) {
         Network network = (Network) person.getGeneCode();
-        double weight = network.getWeightAt(position);
+        float weight = network.getWeightAt(position);
         if (weight + mutationValue > maxStartValue){
             maxStartValue = (int)(weight + mutationValue);
         }
-        network.setWeightAt(position, weight + mutationValue);
+        network.setWeightAt(position, (float) (weight + mutationValue));
     }
 
     @Override
